Restructured the files and directories within RAMPART.
Added first version of ReportGenerator, which is a java app that wraps a simple velocity engine.  Also added a perl script which is intended to run ReportGenerator using a template latex document and a number of variables that will fill in the blanks on the template.

The code that will gather all the variables providing the context for report generator has not been written yet.