Update to V0.11.2.
Fixed a bug with Kmer Genie (new conan release forces clients to specify files for stdout and stderr when process requires it).  RAMPART now specifies logging infor from stderr goes to a file.

Updated ecoli example job config file so that thread counts are reduced.  Also updated the organism element so it relects the recent changes in the interface that V0.11.X expects.

If users are running in an unscheduled environment RAMPART forces all jobs to run linearly.

Starting to incorperate some work on citadel... still work in progress